Refactor tests files
