+++ tests/test_videotools.py
@@ -7,8 +7,11 @@ import sys
 
 import pytest
 
-from moviepy.audio.AudioClip import AudioClip
+from moviepy.audio.AudioClip import AudioClip, CompositeAudioClip
+from moviepy.audio.fx.multiply_volume import multiply_volume
+from moviepy.audio.tools.cuts import find_audio_period
 from moviepy.video.compositing.concatenate import concatenate_videoclips
+from moviepy.video.fx.loop import loop
 from moviepy.video.io.VideoFileClip import VideoFileClip
 from moviepy.video.tools.credits import CreditsClip
 from moviepy.video.tools.cuts import (
@@ -19,7 +22,7 @@ from moviepy.video.tools.cuts import (
 )
 from moviepy.video.VideoClip import BitmapClip, ColorClip, ImageClip, VideoClip
 
-from tests.test_helper import FONT, TMP_DIR, get_stereo_wave
+from tests.test_helper import FONT, TMP_DIR, get_mono_wave, get_stereo_wave
 
 
 try:
@@ -436,6 +439,28 @@ def test_ipython_display_not_available():
     assert str(exc.value) == "Only works inside an IPython Notebook"
 
     del sys.modules["moviepy.video.io.html_tools"]
+
+
+@pytest.mark.parametrize("wave", ("mono", "stereo"))
+def test_find_audio_period(wave):
+    if wave == "mono":
+        wave1 = get_mono_wave(freq=400)
+        wave2 = get_mono_wave(freq=100)
+    else:
+        wave1 = get_stereo_wave(left_freq=400, right_freq=220)
+        wave2 = get_stereo_wave(left_freq=100, right_freq=200)
+    clip = CompositeAudioClip(
+        [
+            AudioClip(make_frame=wave1, duration=0.3, fps=22050),
+            multiply_volume(
+                AudioClip(make_frame=wave2, duration=0.3, fps=22050),
+                0,
+                end_time=0.1,
+            ),
+        ]
+    )
+    loop_clip = loop(clip, 4)
+    assert round(find_audio_period(loop_clip), 6) == pytest.approx(0.29932, 0.1)
 
 
 if __name__ == "__main__":
